review: - rename files - move files to Messenger\Message
