+++ src/main/core/Messenger/ExecuteScheduledTaskHandler.php
@@ -15,10 +15,11 @@ use Claroline\AppBundle\Persistence\ObjectManager;
 use Claroline\CoreBundle\Entity\Task\ScheduledTask;
 use Claroline\CoreBundle\Manager\MailManager;
 use Claroline\CoreBundle\Manager\Task\ScheduledTaskManager;
+use Claroline\CoreBundle\Messenger\Message\ExecuteScheduledTask;
 use Claroline\MessageBundle\Manager\MessageManager;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 
-class ExecuteScheduledTask implements MessageHandlerInterface
+class ExecuteScheduledTaskHandler implements MessageHandlerInterface
 {
     private $messageManager;
     private $taskManager;
@@ -37,7 +38,7 @@ class ExecuteScheduledTask implements MessageHandlerInterface
         $this->objectManager = $objectManager;
     }
 
-    public function __invoke(ScheduledTaskMessage $scheduledTaskMessage)
+    public function __invoke(ExecuteScheduledTask $scheduledTaskMessage)
     {
         $task = $this->objectManager->getRepository(ScheduledTask::class)->find($scheduledTaskMessage->getTaskId());
